+++ src/money/index.ts
@@ -1,0 +1,3 @@
+// Copyright 2026 PeopleWare N.V.
+// SPDX-License-Identifier: Apache-2.0
+export { IbanSchema, type Iban } from "./iban.schema.js";